Improve AppVeyor script.
Exports (and uses) a Powershell function which allows an arbitrary command to be run in 'cmd' with
its stderr redirected to stdout.  This allows cargo commands (most of which output to stderr even in
non-error cases) to be run from an AppVeyor Powershell script without causing AppVeyor to treat the
error output as an indication that the command failed.